Add files via upload
An #python implementation of the generalised Kelly Criterion (a.k.a. The Real Kelly) for outright betting as discussed here https://www.pinnacle.com/en/betting-articles/Betting-Strategy/the-real-kelly-criterion/HZKJTFCB3KNYN9CJ

The algorithm will find optimal bet sizes for a set of mutually exclusive outcomes. This could be a set of futures bets on several different teams (e.g. Liverpool to win the Premier League) or a 3-way betting market (1X2) as explained in example #1 in the article above.